requirements.txt
streamlit
psycopg2
python-dotenv
requests
beautifulsoup4
pandas